+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Release Notes
+
+## 0.9
+* Introduced currencies
+* Introduced languages
+
 ## 0.8
 Refactored using Vue